fillGrasps should call grasp planner even if the object has no type
When we use pick() without grasps, just with the object name, the grasps will be filled with fillGrasps().
What we want is that the external grasp planning service (which we need to implement ourselves) generates the grasps.
This didn't work because fillGrasps required an object type.

This should work without a type, because we can get the position and the shape of the object out of the plenning scene.

This commit doesn't change the behavior but forwards the request to the service even if no type is specified.